Javid minor2/3
Added health bar that slowly depletes over time using a timer. Created 3 scenes for pickups of small, medium and large. Tweaked the enemy to instantiate a random choice between the 3 pick ups (it turns off the visibility and animator after exploding with an animation event, then a timer is set to give player time to grab the power up before ultimately deleting itself